Extract blog date parsing into content-dates utility
